refactor: moved RPC calls to WebAppHandler
